+++ experiments/mirrors_content_analysis_2026_04_24/analysis/__init__.py
@@ -1,5 +1,3 @@
 """Analysis modules for mirrors content analysis experiments."""
 
-from .length_compression_analysis import run_analysis
-
-__all__ = ["run_analysis"]
+__all__: list[str] = []
